Add more tests, change regex
