+++ jest.config.js
@@ -11,7 +11,7 @@ module.exports = {
     "json",
     "node"
   ],
-  testRegex: '(/tests/.*|(\\.|/)(test|spec))\\.(ts|js)x?$',
+  testRegex: '/tests/.*\\.(test|spec)\\.(ts|js)x?$',
   coverageDirectory: 'coverage',
   collectCoverageFrom: [
     'src/**/*.{ts,tsx,js,jsx}',
