+++ tests/Renderer/Camera.spec.ts
@@ -1,58 +1,7 @@
 import { Camera } from '../../src/renderer/Camera';
+import '../glMatrix';
 
 import { mat4, quat, vec3 } from 'gl-matrix';
-
-declare global {
-    namespace jest {
-        interface Matchers<R> {
-            toEqualMat4(argument: mat4): { pass: boolean; message(): string };
-            toEqualVec3(argument: vec3): { pass: boolean; message(): string };
-            toEqualQuat(argument: quat): { pass: boolean; message(): string };
-        }
-    }
-}
-
-expect.extend({
-    toEqualMat4(received: mat4, argument: mat4) {
-        if (mat4.equals(received, argument)) {
-            return {
-                message: () => `expected ${mat4.str(received)} to not be equal to ${mat4.str(argument)}`,
-                pass: true,
-            };
-        } else {
-            return {
-                message: () => `expected ${mat4.str(received)} to be equal to ${mat4.str(argument)}`,
-                pass: false,
-            };
-        }
-    },
-    toEqualVec3(received: vec3, argument: vec3) {
-        if (vec3.equals(received, argument)) {
-            return {
-                message: () => `expected ${vec3.str(received)} to not be equal to ${vec3.str(argument)}`,
-                pass: true,
-            };
-        } else {
-            return {
-                message: () => `expected ${vec3.str(received)} to be equal to ${vec3.str(argument)}`,
-                pass: false,
-            };
-        }
-    },
-    toEqualQuat(received: quat, argument: quat) {
-        if (quat.equals(received, argument)) {
-            return {
-                message: () => `expected ${quat.str(received)} to not be equal to ${quat.str(argument)}`,
-                pass: true,
-            };
-        } else {
-            return {
-                message: () => `expected ${quat.str(received)} to be equal to ${quat.str(argument)}`,
-                pass: false,
-            };
-        }
-    },
-});
 
 describe('Camera', () => {
     it('defaults to an identity transformation', () => {
@@ -89,5 +38,18 @@ describe('Camera', () => {
         camera.moveToWithFixedTarget(vec3.fromValues(0, 1, 1));
         camera.moveByWithFixedTarget(vec3.fromValues(0, 1, 0));
         expect(vec3.transformMat4(vec3.create(), vec3.fromValues(0, 0, -1), camera.getTransform())).toEqualVec3(vec3.fromValues(0, 0, -Math.sqrt(2) * 2));
+    });
+
+    it('can be rotated', () => {
+        const camera = new Camera();
+        camera.setRotation(quat.fromEuler(quat.create(), 0, 90, 0));
+        expect(vec3.transformMat4(vec3.create(), vec3.fromValues(1, 0, 0), camera.getTransform())).toEqualVec3(vec3.fromValues(0, 0, 1));
+    });
+
+    it('can be rotated incrementally', () => {
+        const camera = new Camera();
+        camera.setRotation(quat.fromEuler(quat.create(), 0, 90, 0));
+        camera.rotate(quat.fromEuler(quat.create(), 0, 90, 0));
+        expect(vec3.transformMat4(vec3.create(), vec3.fromValues(0, 0, -1), camera.getTransform())).toEqualVec3(vec3.fromValues(0, 0, 1));
     });
 });
