+++ tslint.json
@@ -16,6 +16,7 @@
     "no-multiline-string": false,
     "no-require-imports": false,
     "no-suspicious-comment": false,
+    "no-import-side-effect": false,
     "typedef": [
       true,
       "parameter",
